fix(BBS-Ruby): fix broken API ver comparison due to floating-point rounding errors
(double)0.111 or 0.111 ≈ 0.11100000000000000144328993201270
(float)0.111 or 0.111f ≈ 0.11100000143051147460937500000000

Therefore, (float)0.111 != 0.111.

I chose to use `double` to solve this issue
because using `float` to solve it would make the code look more bizarre.

* run_ruby():
    * Change the type of local `d` from `float` to `double`

+++ so/bbsruby.c
@@ -1115,7 +1115,7 @@ void run_ruby(
     BRB_C(bbsruby_load_TOC)(cStart, cEnd);
     // Check interface version
     VALUE toc_apiver = CMRB_C(rb_hash_aref, mrb_hash_get)(TOCs_rubyhash, RB_C(rb_str_new_cstr)(TOCs_HEADER[0]));
-    float d = 0;
+    double d = 0; // Prevent further floating-point rounding errors
     if (RTEST(toc_apiver))
         d = atof(StringValueCStr(toc_apiver));
     move(b_lines - 1, 0);
